function/enqueue_request/requirements.txt
azure-functions
azure-identity
azure-storage-queue
